Consolidate test-data generation to single script
Jira ticket: CAMS-135
