+++ dev-tools/test-data/fixtures/reopenedCases.ts
@@ -1,42 +1,14 @@
-import { BCase, Judge, toDbRecords } from '../domain/bcase';
+import { BCase } from '../domain/bcase';
+import { CreateCaseOptions, createCase } from './lib/common';
 
-export function createReopenedCases() {
-  const bCases: Array<BCase> = [];
+export function createReopenedCases(options: CreateCaseOptions) {
+  const decorations: Array<Partial<BCase>> = [];
 
-  const chapter = '15';
-  const county = 'NEW YORK-NY';
-  const courtId = '0208';
-  const group = 'NY';
-  const div = '081';
-  const reopenCode = '1';
-
-  const judge: Judge = {
-    firstName: 'Meyer',
-    lastName: 'Steven',
-  };
-
-  bCases.push({
-    dxtrId: '317344',
-    caseId: '22-66227',
-    shortTitle: 'Smith & Co',
-    chapter,
-    county,
-    group,
-    div,
-    courtId,
-    judge,
-    reopenCode,
+  // Add a number of cases where the case is simply closed.
+  decorations.push({
     transactions: [
       { code: 'CBC', date: '2022-07-27', meta: 'Case Closed' },
       { code: 'OCO', date: '2022-08-27', meta: '41 Order Granting' },
-    ],
-    parties: [
-      {
-        role: 'db',
-        lastName: 'Smith & Co',
-        taxId: '12-1234567',
-        address1: '',
-      },
     ],
     cfValue: 'JtAdm, CLOSED',
     dateFiled: '2022-02-15',
@@ -44,28 +16,10 @@ export function createReopenedCases() {
     dateTerm: '2023-10-01',
   });
 
-  bCases.push({
-    dxtrId: '962282',
-    caseId: '22-75228',
-    shortTitle: 'Johnson Corporation',
-    chapter,
-    county,
-    group,
-    div,
-    courtId,
-    judge,
-    reopenCode,
+  decorations.push({
     transactions: [
       { code: 'CBC', date: '2022-04-27', meta: 'Case Closed' },
       { code: 'OCO', date: '2022-06-14', meta: '41 Order Granting' },
-    ],
-    parties: [
-      {
-        role: 'db',
-        lastName: 'Johnson Corporation',
-        taxId: '00-1234567',
-        address1: '123 Main St',
-      },
     ],
     cfValue: 'JtAdm, CLOSED',
     dateFiled: '2022-02-15',
@@ -73,28 +27,10 @@ export function createReopenedCases() {
     dateTerm: '2023-10-01',
   });
 
-  bCases.push({
-    dxtrId: '639825',
-    caseId: '22-29929',
-    shortTitle: 'Brown Enterprises',
-    chapter,
-    county,
-    group,
-    div,
-    courtId,
-    judge,
-    reopenCode,
+  decorations.push({
     transactions: [
       { code: 'CBC', date: '2022-03-27', meta: 'Case Closed' },
       { code: 'OCO', date: '2022-06-14', meta: '41 Order Granting' },
-    ],
-    parties: [
-      {
-        role: 'db',
-        lastName: 'Brown Enterprises',
-        taxId: '00-1134567',
-        address1: '123 Main St',
-      },
     ],
     cfValue: 'JtAdm, CLOSED',
     dateFiled: '2022-06-20',
@@ -102,28 +38,10 @@ export function createReopenedCases() {
     dateTerm: '2023-10-01',
   });
 
-  bCases.push({
-    dxtrId: '507386',
-    caseId: '22-77102',
-    shortTitle: 'Davis Ltd',
-    chapter,
-    county,
-    group,
-    div,
-    courtId,
-    judge,
-    reopenCode,
+  decorations.push({
     transactions: [
       { code: 'CBC', date: '2022-09-27', meta: 'Case Closed' },
       { code: 'OCO', date: '2022-11-03', meta: '41 Order Granting' },
-    ],
-    parties: [
-      {
-        role: 'db',
-        lastName: 'Davis Ltd',
-        taxId: '45-9876543',
-        address1: '123 Main St',
-      },
     ],
     cfValue: 'MDisCs, CLOSED, Lead',
     dateFiled: '2022-10-04',
@@ -131,28 +49,10 @@ export function createReopenedCases() {
     dateTerm: '2023-10-01',
   });
 
-  bCases.push({
-    dxtrId: '182201',
-    caseId: '22-82448',
-    shortTitle: 'Wilson & Sons',
-    chapter,
-    county,
-    group,
-    div,
-    courtId,
-    judge,
-    reopenCode,
+  decorations.push({
     transactions: [
       { code: 'CBC', date: '2022-01-27', meta: 'Case Closed' },
       { code: 'OCO', date: '2022-11-30', meta: '41 Order Granting' },
-    ],
-    parties: [
-      {
-        role: 'db',
-        lastName: 'Wilson & Sons',
-        taxId: '23-4567890',
-        address1: '123 Main St',
-      },
     ],
     cfValue: 'PENAP, CLOSED',
     dateFiled: '2022-10-11',
@@ -160,29 +60,10 @@ export function createReopenedCases() {
     dateTerm: '2023-10-01',
   });
 
-  bCases.push({
-    dxtrId: '146860',
-    caseId: '22-06498',
-    shortTitle: 'Emily Anderson',
-    chapter,
-    county,
-    group,
-    div,
-    courtId,
-    judge,
-    reopenCode,
+  decorations.push({
     transactions: [
       { code: 'CBC', date: '2022-08-20', meta: 'Case Closed' },
       { code: 'OCO', date: '2022-12-31', meta: '41 Order Granting' },
-    ],
-    parties: [
-      {
-        role: 'db',
-        lastName: 'Anderson',
-        firstName: 'Emily',
-        ssn: '001-23-4567',
-        address1: '123 Main St',
-      },
     ],
     cfValue: 'JtAdm, CGM1, CLOSED',
     dateFiled: '2022-10-18',
@@ -191,32 +72,13 @@ export function createReopenedCases() {
   });
 
   // Example of a case with multiple closure and reopen transactions ending with a DISMISSAL.
-  bCases.push({
-    dxtrId: '506663',
-    caseId: '22-86620',
-    shortTitle: 'Benjamin Wilson',
-    chapter,
-    county,
-    group,
-    div,
-    courtId,
-    judge,
-    reopenCode,
+  decorations.push({
     transactions: [
       { code: 'CBC', date: '2022-03-14', meta: 'Case Closed' },
       { code: 'OCO', date: '2023-03-13', meta: '41 Order Granting' },
       { code: 'CBC', date: '2022-03-14', meta: 'Case Closed' },
       { code: 'OCO', date: '2023-03-13', meta: '41 Order Granting' },
       { code: 'CDC', date: '2023-11-01', meta: 'Case Dismissed' },
-    ],
-    parties: [
-      {
-        role: 'db',
-        lastName: 'Wilson',
-        firstName: 'Benjamin',
-        ssn: '123-45-6789',
-        address1: '123 Main St',
-      },
     ],
     cfValue: 'FeeDueAP, PENAP, Lead, Mediation',
     dateFiled: '2022-12-19',
@@ -226,30 +88,11 @@ export function createReopenedCases() {
   });
 
   // Example of a reopened case followed by a CLOSURE.
-  bCases.push({
-    dxtrId: '882356',
-    caseId: '23-70851',
-    shortTitle: 'Sophia Martin',
-    chapter,
-    county,
-    group,
-    div,
-    courtId,
-    judge,
-    reopenCode,
+  decorations.push({
     transactions: [
       { code: 'CBC', date: '2023-01-25', meta: 'Case Closed' },
       { code: 'OCO', date: '2023-09-14', meta: '41 Order Granting' },
       { code: 'CBC', date: '2023-10-01', meta: 'Case Closed' },
-    ],
-    parties: [
-      {
-        role: 'db',
-        lastName: 'Martin',
-        firstName: 'Sophia',
-        ssn: '987-65-4321',
-        address1: '123 Main St',
-      },
     ],
     cfValue: 'CLOSED',
     dateFiled: '2023-04-06',
@@ -258,31 +101,12 @@ export function createReopenedCases() {
   });
 
   // Example of a case with multiple closures and reopen transactions ending with a REOPEN.
-  bCases.push({
-    dxtrId: '449865',
-    caseId: '23-07045',
-    shortTitle: 'Ethan Johnson',
-    chapter,
-    county,
-    group,
-    div,
-    courtId,
-    judge,
-    reopenCode,
+  decorations.push({
     transactions: [
       { code: 'CBC', date: '2023-07-08', meta: 'Case Closed' },
       { code: 'OCO', date: '2023-08-15', meta: '41 Order Granting' },
       { code: 'CBC', date: '2023-09-09', meta: 'Case Closed' },
       { code: 'OCO', date: '2023-10-04', meta: '41 Order Granting' },
-    ],
-    parties: [
-      {
-        role: 'db',
-        lastName: 'Johnson',
-        firstName: 'Ethan',
-        taxId: '00-1134567',
-        address1: '123 Main St',
-      },
     ],
     cfValue: 'CLOSED',
     dateFiled: '2023-06-28',
@@ -291,29 +115,11 @@ export function createReopenedCases() {
   });
 
   // Example of a reopened case followed by a DISMISSAL.
-  bCases.push({
-    dxtrId: '129204',
-    caseId: '23-16111',
-    shortTitle: 'Olivia Davis',
-    chapter,
-    county,
-    group,
-    div,
-    courtId,
-    judge,
-    reopenCode,
+  decorations.push({
     transactions: [
       { code: 'CBC', date: '2023-06-01', meta: 'Case Closed' },
       { code: 'OCO', date: '2023-10-23', meta: '41 Order Granting' },
       { code: 'CDC', date: '2023-11-01', meta: 'Case Dismissed' },
-    ],
-    parties: [
-      {
-        role: 'db',
-        lastName: 'Davis',
-        firstName: 'Olivia',
-        address1: '123 Main St',
-      },
     ],
     cfValue: 'CLOSED',
     dateFiled: '2023-07-17',
@@ -322,5 +128,10 @@ export function createReopenedCases() {
     dateDismiss: '2023-11-01',
   });
 
-  return toDbRecords(bCases);
+  // Now create some cases and append/override properties with the decorations.
+  const bCases: Array<BCase> = decorations.map((decoration) => {
+    return { ...createCase(options), ...decoration };
+  });
+
+  return bCases;
 }
