+++ dev-tools/test-data/utility.ts
@@ -1,4 +1,3 @@
-import { Faker, fakerEN_GB, fakerEN_US, fakerES_MX } from '@faker-js/faker';
 import { ColumnNames, TableRecordHelper } from './types';
 
 /**
@@ -106,22 +105,6 @@ export function randomTruth() {
 
 export function randomInt(range: number) {
   return Math.floor(Math.random() * range);
-}
-
-export function getFakerLocale(): Faker {
-  const locales = [
-    { faker: fakerEN_US, countryCode: 'US', countryName: 'United States' },
-    { faker: fakerEN_GB, countryCode: 'UK', countryName: 'United Kingdom' },
-    { faker: fakerES_MX, countryCode: 'MX', countryName: 'Mexico' },
-  ];
-  const { faker, countryCode, countryName } = locales[randomInt(locales.length)];
-  faker.location.countryCode = () => {
-    return countryCode;
-  };
-  faker.location.country = () => {
-    return countryName;
-  };
-  return faker;
 }
 
 export function removeExtraSpaces(s: string | undefined): string | undefined {
